+++ encore_api_cli/commands/movie.py
@@ -30,7 +30,7 @@ def show(state: State, movie_id: int) -> None:
     client = get_client(state)
 
     try:
-        data = client.get_one_data("movies", movie_id)
+        data = client.get_movie(movie_id)
     except RequestsError as e:
         raise ClickException(str(e))
 
@@ -47,9 +47,9 @@ def list(state: State) -> None:
     try:
         if state.use_spinner:
             with yaspin(text="Retrieving..."):
-                data = client.get_list_data("movies")
+                data = client.get_movies()
         else:
-            data = client.get_list_data("movies")
+            data = client.get_movies()
     except RequestsError as e:
         raise ClickException(str(e))