+++ tests/unit/commands/test_analysis.py
@@ -13,6 +13,94 @@ def test_analysis(runner):
 
 class TestAnalysisShow(object):
     @pytest.mark.parametrize(
+        "args, status, expected",
+        [
+            (
+                ["analysis", "show", "1"],
+                "SUCCESS",
+                dedent(
+                    """\
+
+                    {
+                      "id": 111,
+                      "result": [
+                        [
+                          {
+                            "type": "angle",
+                            "values": [
+                              180
+                            ]
+                          }
+                        ]
+                      ],
+                      "execStatus": "SUCCESS"
+                    }
+
+                    """
+                ),
+            ),
+            (
+                ["analysis", "show", "1"],
+                "FAILURE",
+                dedent(
+                    """\
+
+                    {
+                      "id": 111,
+                      "result": null,
+                      "execStatus": "FAILURE"
+                    }
+
+                    """
+                ),
+            ),
+            (
+                ["analysis", "show", "1", "--no-result"],
+                "SUCCESS",
+                dedent(
+                    """\
+
+                    {
+                      "id": 111,
+                      "execStatus": "SUCCESS"
+                    }
+
+                    """
+                ),
+            ),
+            (
+                ["analysis", "show", "1", "--no-result"],
+                "FAILURE",
+                dedent(
+                    """\
+
+                    {
+                      "id": 111,
+                      "execStatus": "FAILURE"
+                    }
+
+                    """
+                ),
+            ),
+        ],
+    )
+    def test_valid(self, runner, make_client, args, status, expected):
+        client_mock = make_client(status)
+        result = runner.invoke(cli, args)
+
+        assert client_mock.call_count == 1
+        assert result.exit_code == 0
+        assert result.output == expected
+
+    @pytest.mark.parametrize(
+        "args",
+        [
+            ["analysis", "show", "1", "--only"],
+            ["analysis", "show", "1", "--only-result"],
+            ["analysis", "show", "1", "--only", "--only-result"],
+        ],
+    )
+    @pytest.mark.parametrize(
         "status, expected",
         [
             (
@@ -37,9 +125,9 @@ class TestAnalysisShow(object):
             ("FAILURE", "Error: Status is not SUCCESS.\n"),
         ],
     )
-    def test_valid(self, mocker, runner, status, expected):
-        client_mock = self._get_client_mock(mocker, status)
-        result = runner.invoke(cli, ["analysis", "show", "1"])
+    def test_with_only(self, runner, make_client, args, status, expected):
+        client_mock = make_client(status)
+        result = runner.invoke(cli, args)
 
         assert client_mock.call_count == 1
         assert result.exit_code == 0
@@ -56,37 +144,58 @@ class TestAnalysisShow(object):
                     "invalid_id is not a valid integer\n",
                 ),
             ),
+            (
+                ["analysis", "show", "1", "--only", "--no-result"],
+                (
+                    'Error: "--only, --only-result" and "--no-result" options '
+                    "cannot be used at the same time.\n"
+                ),
+            ),
+            (
+                ["analysis", "show", "1", "--only-result", "--no-result"],
+                (
+                    'Error: "--only, --only-result" and "--no-result" options '
+                    "cannot be used at the same time.\n"
+                ),
+            ),
         ],
     )
-    def test_invalid_params(self, runner, args, expected):
+    def test_invalid_params(self, runner, make_client, args, expected):
+        client_mock = make_client()
         result = runner.invoke(cli, args)
+
+        assert client_mock.call_count == 0
         assert result.exit_code == 2
         assert result.output.endswith(expected)
 
-    def test_with_error(self, mocker, runner):
-        client_mock = self._get_client_mock(mocker, with_exception=True)
+    def test_with_error(self, runner, make_client):
+        client_mock = make_client(with_exception=True)
         result = runner.invoke(cli, ["analysis", "show", "1"])
 
         assert client_mock.call_count == 1
         assert result.exit_code == 1
-        assert "Error" in result.output
+        assert result.output == "Error: \n"
 
-    def _get_client_mock(self, mocker, status="SUCCESS", with_exception=False):
-        client_mock = mocker.MagicMock()
-        if status == "SUCCESS":
-            data = [[{"type": "angle", "values": [180]}]]
-        else:
-            data = None
-        if with_exception:
-            client_mock.return_value.get_one_data.side_effect = RequestsError()
-        else:
-            client_mock.return_value.get_one_data.return_value = {
-                "id": 111,
-                "result": data,
-                "execStatus": status,
-            }
-        mocker.patch("encore_api_cli.commands.analysis.get_client", client_mock)
-        return client_mock
+    @pytest.fixture
+    def make_client(self, mocker):
+        def _make_client(status="SUCCESS", with_exception=False):
+            client_mock = mocker.MagicMock()
+            if status == "SUCCESS":
+                data = [[{"type": "angle", "values": [180]}]]
+            else:
+                data = None
+            if with_exception:
+                client_mock.return_value.get_analysis.side_effect = RequestsError()
+            else:
+                client_mock.return_value.get_analysis.return_value = {
+                    "id": 111,
+                    "result": data,
+                    "execStatus": status,
+                }
+            mocker.patch("encore_api_cli.commands.analysis.get_client", client_mock)
+            return client_mock
+
+        return _make_client
 
 
 class TestAnalysisList(object):
@@ -98,8 +207,8 @@ class TestAnalysisList(object):
             ["analysis", "list", "--status", "success"],
         ],
     )
-    def test_valid(self, mocker, runner, args):
-        client_mock = self._get_client_mock(mocker)
+    def test_valid(self, runner, make_client, args):
+        client_mock = make_client()
         expected = dedent(
             """\
 
@@ -119,9 +228,9 @@ class TestAnalysisList(object):
         assert result.exit_code == 0
         assert result.output == expected
 
-    def test_with_spinner(self, mocker, monkeypatch, runner):
+    def test_with_spinner(self, monkeypatch, runner, make_client):
         monkeypatch.setenv("ANYMOTION_USE_SPINNER", "true")
-        client_mock = self._get_client_mock(mocker)
+        client_mock = make_client()
 
         result = runner.invoke(cli, ["analysis", "list"])
 
@@ -129,8 +238,8 @@ class TestAnalysisList(object):
         assert result.exit_code == 0
         assert "Retrieving..." in result.output
 
-    def test_with_pager(self, mocker, runner):
-        client_mock = self._get_client_mock(mocker, num_data=10)
+    def test_with_pager(self, runner, make_client):
+        client_mock = make_client(num_data=10)
 
         result = runner.invoke(cli, ["analysis", "list"])
 
@@ -155,28 +264,32 @@ class TestAnalysisList(object):
             ),
         ],
     )
-    def test_invalid_params(self, mocker, runner, args, expected):
-        client_mock = self._get_client_mock(mocker)
+    def test_invalid_params(self, runner, make_client, args, expected):
+        client_mock = make_client()
         result = runner.invoke(cli, args)
 
         assert client_mock.call_count == 0
         assert result.exit_code == 2
         assert result.output.endswith(expected)
 
-    def test_with_error(self, mocker, runner):
-        client_mock = self._get_client_mock(mocker, with_exception=True)
+    def test_with_error(self, runner, make_client):
+        client_mock = make_client(with_exception=True)
         result = runner.invoke(cli, ["analysis", "list"])
 
         assert client_mock.call_count == 1
         assert result.exit_code == 1
         assert "Error" in result.output
 
-    def _get_client_mock(self, mocker, num_data=1, with_exception=False):
-        client_mock = mocker.MagicMock()
-        if with_exception:
-            client_mock.return_value.get_list_data.side_effect = RequestsError()
-        else:
-            data = [{"id": i + 1, "execStatus": "SUCCESS"} for i in range(num_data)]
-            client_mock.return_value.get_list_data.return_value = data
-        mocker.patch("encore_api_cli.commands.analysis.get_client", client_mock)
-        return client_mock
+    @pytest.fixture
+    def make_client(self, mocker):
+        def _make_client(num_data=1, with_exception=False):
+            client_mock = mocker.MagicMock()
+            if with_exception:
+                client_mock.return_value.get_analyses.side_effect = RequestsError()
+            else:
+                data = [{"id": i + 1, "execStatus": "SUCCESS"} for i in range(num_data)]
+                client_mock.return_value.get_analyses.return_value = data
+            mocker.patch("encore_api_cli.commands.analysis.get_client", client_mock)
+            return client_mock
+
+        return _make_client