+++ tests/unit/commands/test_keypoint.py
@@ -12,34 +12,121 @@ def test_keypoint(runner):
 
 
 class TestKeypointShow(object):
-    def test_valid(self, mocker, runner):
-        client_mock = self._get_client_mock(mocker)
-        result = runner.invoke(cli, ["keypoint", "show", "1"])
+    @pytest.mark.parametrize(
+        "args, status, expected",
+        [
+            (
+                ["keypoint", "show", "111"],
+                "SUCCESS",
+                dedent(
+                    """\
+
+                    {
+                      "id": 111,
+                      "keypoint": [
+                        {
+                          "1": [
+                            1,
+                            0
+                          ]
+                        }
+                      ],
+                      "execStatus": "SUCCESS"
+                    }
+
+                    """
+                ),
+            ),
+            (
+                ["keypoint", "show", "111"],
+                "FAILURE",
+                dedent(
+                    """\
+
+                    {
+                      "id": 111,
+                      "keypoint": null,
+                      "execStatus": "FAILURE"
+                    }
+
+                    """
+                ),
+            ),
+            (
+                ["keypoint", "show", "111", "--no-keypoint"],
+                "SUCCESS",
+                dedent(
+                    """\
+
+                    {
+                      "id": 111,
+                      "execStatus": "SUCCESS"
+                    }
+
+                    """
+                ),
+            ),
+            (
+                ["keypoint", "show", "111", "--no-keypoint"],
+                "FAILURE",
+                dedent(
+                    """\
+
+                    {
+                      "id": 111,
+                      "execStatus": "FAILURE"
+                    }
+
+                    """
+                ),
+            ),
+        ],
+    )
+    def test_valid(self, runner, make_client, args, status, expected):
+        client_mock = make_client(status)
+        result = runner.invoke(cli, args)
 
         assert client_mock.call_count == 1
         assert result.exit_code == 0
-        assert result.output == dedent(
-            """\
+        assert result.output == expected
 
-                [
-                  {
-                    "1": [
-                      1,
-                      0
+    @pytest.mark.parametrize(
+        "args",
+        [
+            ["keypoint", "show", "1", "--only"],
+            ["keypoint", "show", "1", "--only-keypoint"],
+        ],
+    )
+    @pytest.mark.parametrize(
+        "status, expected",
+        [
+            (
+                "SUCCESS",
+                dedent(
+                    """\
+
+                    [
+                      {
+                        "1": [
+                          1,
+                          0
+                        ]
+                      }
                     ]
-                  }
-                ]
 
-            """
-        )
-
-    def test_valid_not_success(self, mocker, runner):
-        client_mock = self._get_client_mock(mocker, status="FAILURE")
-        result = runner.invoke(cli, ["keypoint", "show", "1"])
+                    """
+                ),
+            ),
+            ("FAILURE", "Error: Status is not SUCCESS.\n"),
+        ],
+    )
+    def test_with_only(self, runner, make_client, args, status, expected):
+        client_mock = make_client(status)
+        result = runner.invoke(cli, args)
 
         assert client_mock.call_count == 1
         assert result.exit_code == 0
-        assert result.output == "Error: Status is not SUCCESS.\n"
+        assert result.output == expected
 
     @pytest.mark.parametrize(
         "args, expected",
@@ -52,51 +139,74 @@ class TestKeypointShow(object):
                     "invalid_id is not a valid integer\n"
                 ),
             ),
+            (
+                ["keypoint", "show", "1", "--only", "--no-keypoint"],
+                (
+                    'Error: "--only, --only-keypoint" and "--no-keypoint" options '
+                    "cannot be used at the same time.\n"
+                ),
+            ),
+            (
+                ["keypoint", "show", "1", "--only-keypoint", "--no-keypoint"],
+                (
+                    'Error: "--only, --only-keypoint" and "--no-keypoint" options '
+                    "cannot be used at the same time.\n"
+                ),
+            ),
         ],
     )
-    def test_invalid_params(self, mocker, runner, args, expected):
-        client_mock = self._get_client_mock(mocker)
+    def test_invalid_params(self, runner, make_client, args, expected):
+        client_mock = make_client()
         result = runner.invoke(cli, args)
 
         assert client_mock.call_count == 0
         assert result.exit_code == 2
         assert result.output.endswith(expected)
 
-    def test_with_pager(self, mocker, runner):
-        client_mock = self._get_client_mock(mocker, num_data=10)
+    @pytest.mark.parametrize(
+        "args",
+        [
+            ["keypoint", "show", "1", "--only"],
+            ["keypoint", "show", "1", "--only-keypoint"],
+        ],
+    )
+    def test_with_pager(self, runner, make_client, args):
+        client_mock = make_client(num_data=10)
 
-        result = runner.invoke(cli, ["keypoint", "show", "1"])
+        result = runner.invoke(cli, args)
 
         assert client_mock.call_count == 1
         assert result.exit_code == 0
         assert '"1": [\n      10,\n      81\n    ]\n' in result.output
 
-    def test_with_error(self, mocker, runner):
-        client_mock = self._get_client_mock(mocker, with_exception=True)
+    def test_with_error(self, runner, make_client):
+        client_mock = make_client(with_exception=True)
         result = runner.invoke(cli, ["keypoint", "show", "1"])
 
         assert client_mock.call_count == 1
         assert result.exit_code == 1
-        assert "Error" in result.output
+        assert result.output == "Error: \n"
 
-    def _get_client_mock(
-        self, mocker, status="SUCCESS", num_data=1, with_exception=False
-    ):
-        if status == "SUCCESS":
-            data = [{"1": [i + 1, i * i]} for i in range(num_data)]
-        else:
-            data = None
-        client_mock = mocker.MagicMock()
-        if with_exception:
-            client_mock.return_value.get_one_data.side_effect = RequestsError()
-        else:
-            client_mock.return_value.get_one_data.return_value = {
-                "id": 111,
-                "keypoint": data,
-                "execStatus": status,
-            }
-        mocker.patch("encore_api_cli.commands.keypoint.get_client", client_mock)
-        return client_mock
+    @pytest.fixture
+    def make_client(self, mocker):
+        def _make_client(status="SUCCESS", num_data=1, with_exception=False):
+            if status == "SUCCESS":
+                data = [{"1": [i + 1, i * i]} for i in range(num_data)]
+            else:
+                data = None
+            client_mock = mocker.MagicMock()
+            if with_exception:
+                client_mock.return_value.get_keypoint.side_effect = RequestsError()
+            else:
+                client_mock.return_value.get_keypoint.return_value = {
+                    "id": 111,
+                    "keypoint": data,
+                    "execStatus": status,
+                }
+            mocker.patch("encore_api_cli.commands.keypoint.get_client", client_mock)
+            return client_mock
+
+        return _make_client
 
 
 class TestKeypointList(object):
@@ -108,8 +218,8 @@ class TestKeypointList(object):
             ["keypoint", "list", "--status", "success"],
         ],
     )
-    def test_valid(self, mocker, runner, args):
-        client_mock = self._get_client_mock(mocker)
+    def test_valid(self, runner, make_client, args):
+        client_mock = make_client()
         result = runner.invoke(cli, args)
 
         assert client_mock.call_count == 1
@@ -129,9 +239,9 @@ class TestKeypointList(object):
             """
         )
 
-    def test_with_spinner(self, mocker, monkeypatch, runner):
+    def test_with_spinner(self, monkeypatch, runner, make_client):
         monkeypatch.setenv("ANYMOTION_USE_SPINNER", "true")
-        client_mock = self._get_client_mock(mocker)
+        client_mock = make_client()
 
         result = runner.invoke(cli, ["keypoint", "list"])
 
@@ -139,8 +249,8 @@ class TestKeypointList(object):
         assert result.exit_code == 0
         assert "Retrieving..." in result.output
 
-    def test_with_pager(self, mocker, runner):
-        client_mock = self._get_client_mock(mocker, num_data=10)
+    def test_with_pager(self, runner, make_client):
+        client_mock = make_client(num_data=10)
 
         result = runner.invoke(cli, ["keypoint", "list"])
 
@@ -165,31 +275,35 @@ class TestKeypointList(object):
             ),
         ],
     )
-    def test_invalid_params(self, mocker, runner, args, expected):
-        client_mock = self._get_client_mock(mocker)
+    def test_invalid_params(self, runner, make_client, args, expected):
+        client_mock = make_client()
         result = runner.invoke(cli, args)
 
         assert client_mock.call_count == 0
         assert result.exit_code == 2
         assert result.output.endswith(expected)
 
-    def test_with_error(self, mocker, runner):
-        client_mock = self._get_client_mock(mocker, with_exception=True)
+    def test_with_error(self, runner, make_client):
+        client_mock = make_client(with_exception=True)
         result = runner.invoke(cli, ["keypoint", "list"])
 
         assert client_mock.call_count == 1
         assert result.exit_code == 1
         assert "Error" in result.output
 
-    def _get_client_mock(self, mocker, num_data=1, with_exception=False):
-        client_mock = mocker.MagicMock()
-        if with_exception:
-            client_mock.return_value.get_list_data.side_effect = RequestsError()
-        else:
-            data = [
-                {"id": i + 1, "image": 2, "movie": None, "execStatus": "SUCCESS"}
-                for i in range(num_data)
-            ]
-            client_mock.return_value.get_list_data.return_value = data
-        mocker.patch("encore_api_cli.commands.keypoint.get_client", client_mock)
-        return client_mock
+    @pytest.fixture
+    def make_client(self, mocker):
+        def _make_client(num_data=1, with_exception=False):
+            client_mock = mocker.MagicMock()
+            if with_exception:
+                client_mock.return_value.get_keypoints.side_effect = RequestsError()
+            else:
+                data = [
+                    {"id": i + 1, "image": 2, "movie": None, "execStatus": "SUCCESS"}
+                    for i in range(num_data)
+                ]
+                client_mock.return_value.get_keypoints.return_value = data
+            mocker.patch("encore_api_cli.commands.keypoint.get_client", client_mock)
+            return client_mock
+
+        return _make_client