+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Unreleased
 
+- Added `--only` option to `keypoint show` and `analysis show` commands.
 - Moved sdk to another package [encore_sdk](https://bitbucket.org/nttpc-datascience/encore-sdk).
 
 ## 0.8.0